protocol: fix endianness of transport_lenght field
The USB device failed to respond and was triggering timeout errors due
to the host sending an incorrect transport length. Changing endianness
for this field fixed the issue.

According to the USB Mass Storage Class specification: "All CBW
transfers shall be ordered with the LSB (byte 0) first (little endian)."

This error seems to be systematically present throughout the
serialization of command blocks. Luckily, these incidents are immune to
endianness changes in a way that does not affect the resulting behavior.
This does not take away that the code in its current form is
error-prone.

Be aware that '.put_uX(..)' will append values Big-endian.
'.put_uX_le(..)' needs to be explicitly called for the data to be parsed
with the correct endianness.